+++ platforms/cuda/include/CudaParallelKernels.h
@@ -84,7 +84,7 @@ private:
     std::vector<Kernel> kernels;
     std::vector<long long> completionTimes;
     std::vector<double> contextNonbondedFractions;
-    int2* interactionCounts;
+    long2* interactionCounts;
     CudaArray contextForces;
     void* pinnedPositionBuffer;
     long long* pinnedForceBuffer;
